backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.3.0
peewee>=3.16.0
python-dotenv>=1.0.0
docker>=6.1.3
psutil>=5.9.0
pytest>=7.4.0
httpx>=0.24.0
